Reflection-Agent/pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["teruyo14 <teruyo.sui4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain = "^0.3.2"
langchain-openai = "^0.2.2"
langgraph = "^0.2.34"
openai = "^1.51.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
